Refactorización Vista Técnico y Cambios de Información en la Página de Cursos
Se refactorizó el contenido de toda la vista de Técnico. Se cambió la información desplegada en la tabla de evaluaciones. Se cambió la información de la tabla de progreso de la página de cursos. Se cambió la UI de la página de cursos para indicar de mejor manera que lecciones han sido completadas y cuales no.
